Create RX module + other small stylistic changes
